% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FHM.R
\name{gridFastHike}
\alias{gridFastHike}
\title{Runs the grid version of the Modified Fast Hiking Method}
\usage{
gridFastHike(domain, seeds, spatial.res = 1, fun = "Tobler")
}
\arguments{
\item{domain}{Grid (matrix) of chosen dimension with diffusivity values
for every grid cell. Values above 1 will boost diffusivity, below 1 will
inhibit it. Values of 0 should mark cells that block diffusion.}

\item{seeds}{A (4 x n) array containing the x-coordinate, y-coordinate,
incept time and rate-of-spread for each of the n seeds.}

\item{spatial.res}{(Optional) Spatial resolution of the grid in metre.
See example below. Defaults to 1 metre.}

\item{fun}{(Optional) Hiking function to use, See \link{hiking.speed} for details.}
}
\description{
This function runs the Modified Fast Hiking Method on a gridded domain.
Output arrival time is in hours.
}
\examples{
# Single process
grid <- matrix(1,10,10)
seed <- data.frame(row=5, col=5, incept=0)
fm <- gridFastHike(grid, seed)
image(fm$arrival.time)

# Two processes with same incept time
seeds <- data.frame(row=c(7,2), col=c(7,2), incept=c(0,0))
fm2 <- gridFastHike(grid, seeds)
par(mfrow=c(1,3))
image(fm2$process, main='process')
image(fm2$arrival.time, main='arrival time')
image(fm2$cost.distance, main='distance')

# Same as before but changing spatial.res parameter
fm3 <- gridFastHike(grid, seeds, spatial.res=10)

# Same as before but with a barrier in middle
grid[5,2:9] <- 0
fm4 <- gridFastHike(grid, seeds, spatial.res=10)
par(mfrow=c(1,3))
image(fm4$process, main='process')
image(fm4$arrival.time, main='arrival time')
image(fm4$cost.distance, main='distance')

# Same as before but with different incept times and speeds
seeds <- data.frame(row=c(7,2), col=c(7,2), incept=c(0,1), v0=c(2,5))
fm5 <- gridFastHike(grid, seeds, spatial.res=10)
par(mfrow=c(1,3))
image(fm5$process, main='process')
image(fm5$arrival.time, main='arrival time')
image(fm5$cost.distance, main='distance')
}
