% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sPath.R
\name{gridSpath}
\alias{gridSpath}
\title{Shortest Path on a gridded domain}
\usage{
gridSpath(cdist, drop, spatial.res = 1)
}
\arguments{
\item{spatial.res}{(Optional) Spatial resolution of the grid in metres.}
}
\description{
TODO description
}
\examples{

}
