% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMM.R
\name{spFastMarch}
\alias{spFastMarch}
\title{Runs the spatial version of the Modified Fast Marching Method}
\usage{
spFastMarch(domain, seeds)
}
\arguments{
\item{domain}{A \code{\link[raster]{raster}} object of chosen dimension
and resolution with diffusivity values for every cell. Values above 1 will
 boost diffusivity, below 1 will inhibit it. Values of 0 should mark cells
 that block diffusion.}

\item{seeds}{A \code{\link[sp]{SpatialPointsDataFrame}} object containing
the incept time and rate-of-spread for each of the n seeds in its data.frame,
 in columns named exactly \emph{incept} (for incept time) and \emph{speed} (
 for rate-of-spread).
This object will be automatically transformed to the projection of \emph{domain}.}
}
\description{
This function runs the Modified Fast Marching Method of Silva and Steele
(2012,2014) from \emph{sp} and \emph{raster} objects and outputs results
in the same formats, makin it more convenient for (geo)spatial analyses
and simulation.
}
\examples{
\dontrun{
domain <- r # a previously constructed raster
coords <- cbind(c(-3, 20), c(37, 40)) # WGS84 Coordinates (Lon and Lat) for seeds
seed.df <- data.frame(incept=c(0,500), speed=c(1,1)) # Incept time and speed for each seed
seeds <- SpatialPointsDataFrame(coords, seed.df, proj4string=CRS("+proj=longlat +datum=WGS84"))

fm <- spFastMarch(domain, seeds)
plot(fm$T)
}
}
\references{
Sethian, J.A. (1996), A fast marching level set method for
monotonically advancing fronts, \emph{Proc. Natl. Acad. Sci.} 93 (4),
1591-1595, doi:

Silva, F. and Steele, J. (2012), Modeling Boundaries Between
Converging Fronts in Prehistory, \emph{Advances in Complex Systems},
 15(1-2), 1150005, doi: 10.1142/S0219525911003293

Silva, F. and Steele, J. (2014), New methods for reconstructing
geographical effects on dispersal rates and routes from large-scale
radiocarbon databases, \emph{Journal of Archaeological Science} 52,
609-620, doi: 10.1016/j.jas.2014.04.021
}
