% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sPath.R
\name{spSpath}
\alias{spSpath}
\title{Shortest Path on a spatial domain}
\usage{
spSpath(surface, droplet)
}
\arguments{
\item{surface}{A \code{\link[raster]{raster}} object with arrival times
or \emph{fastmaRching} object.}

\item{droplet}{A \code{\link[sp]{SpatialPointsDataFrame}} object with
droplet location.}
}
\description{
This function returns the shortest-path from \emph{droplet} location to
source of dispesal surface.
}
