% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sPath.R
\name{gridSpath}
\alias{gridSpath}
\title{Shortest Path on a gridded domain}
\usage{
gridSpath(surface, droplet, spatial.res)
}
\arguments{
\item{surface}{Grid (matrix) of arrival times or \emph{fastmaRching} object.}

\item{droplet}{Location of droplet}

\item{spatial.res}{(Optional) Spatial resolution of the grid in metre.
See example below.}
}
\description{
This function returns the shortest-path from \emph{droplet} location to
source of dispesal surface.
}
