% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FHM.R
\name{spFastHike}
\alias{spFastHike}
\title{Modified Fast Hiking Method on a spatial domain}
\usage{
spFastHike(dem, seeds, spatial.res, fun = "Tobler")
}
\arguments{
\item{dem}{A \code{\link[raster]{raster}} object of chosen dimension
and resolution with elevation data.}

\item{seeds}{A \code{\link[sp]{SpatialPointsDataFrame}} object containing
the incept time and rate-of-spread for each of the n seeds in its data.frame,
 in a colums named exactly \emph{incept} (for incept time), \emph{off.path}
 (see Tobler's function).
This object will be automatically transformed to the projection of \emph{domain}.}

\item{spatial.res}{(Optional) Spatial resolution of the raster in metres.
Defaults to that of the raster used for DEM}

\item{fun}{(Optional) Hiking function to use, See \link{hiking.speed} for details.}
}
\description{
This function runs the Modified Fast Hiking Method from \emph{sp} and
\emph{raster} objects and outputs results in the same formats, making
it more convenient for (geo)spatial analyses and simulation. Output
arrival time is in hours.
}
\examples{
library(raster); library(sp); library(rgdal)
domain <- raster(system.file("external/test.grd", package="raster")) # sample raster
coords <- cbind(c(179000,181200), c(330000, 333000)) # coordinates for seeds
seed.df <- data.frame(incept=c(0,1)) # parameters for each seed
seeds <- SpatialPointsDataFrame(coords, seed.df, proj4string=crs(domain))

fm <- spFastHike(domain, seeds)
}
