% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sPath.R
\name{spSpath}
\alias{spSpath}
\title{Shortest Path on a spatial domain}
\usage{
spSpath(cdist, drop)
}
\arguments{
\item{spatial.res}{(Optional) Spatial resolution of the raster in metres.}
}
\description{
TODO description
}
\examples{

}
